% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldata.R
\docType{data}
\name{coldata}
\alias{coldata}
\title{Sample descriptions for 399 plasma samples}
\format{
A dataframe with 399 rows and 2 columns:
\describe{
  \item{Sample}{Name of the sample}
  \item{disease}{Disease status of patient sample was derived from: Healthy, Disease1, Disease2}
  ...
}
}
\usage{
coldata
}
\description{
A dataset containing the sample names and disease status for 399 plasma samples
}
\keyword{datasets}
