% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites.R
\docType{data}
\name{metabolites}
\alias{metabolites}
\title{Normalised abundance data for 61 metabolites}
\format{
A numeric matrix with 61 rows and 399 columns
}
\usage{
metabolites
}
\description{
A dataset containing the normalised abundance of 61 metabolites detected
in 399 human plasma samples
}
\keyword{datasets}
