% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Generate a volcano plot for an lmFit contrast}
\usage{
volcano_plot(fit, contrast, ref_line = TRUE, title)
}
\arguments{
\item{fit}{An MArrayLM object produced by limma::lmFit or limma::contrasts.fit}

\item{contrast}{String indicating the column name/contrast of interest}

\item{ref_line}{If TRUE, add a reference line for the -log10 adjusted p-value corresponding \cr
to an adjusted p-value of 0.05}

\item{title}{String assigned as the title of the volcano plot}
}
\value{

}
\description{
Generate a volcano plot for an lmFit contrast
}
\examples{
design <- model.matrix(~ 0 + disease, coldata)

colnames(design) <- c("Healthy", "Disease1", "Disease2")

cont.matrix <- limma::makeContrasts("HealthyvsDisease1" = Disease1-Healthy,
                                    "HealthyvsDisease2" = Disease2-Healthy,
                                    "Disease1vsDisease2" = Disease2-Disease1,
                                    levels = design)

fit <- limma::lmFit(metabolites, design = design)

fit2 <- limma::contrasts.fit(fit, cont.matrix)

fit2 <- limma::eBayes(fit2)

volcano_plot(fit2, contrast = "HealthyvsDisease1", title = "Volcano Plot: Healthy vs. Disease 1")
}
